keep all inline efficiency unchanged
